chore: fmt

+++ src/MEVController.sol
@@ -8,20 +8,18 @@ import {MEVTaxedOval} from "./MEVTaxedOval.sol";
  * @title MEVController update the BaseController and remove the need of setting and checking unlockers.
  */
 abstract contract MEVController is BaseController, MEVTaxedOval {
-  
     ///@dev The MEV tax multiplier, people who submit the tx pays msg.fee * mevTaxMultiplier
     uint256 private mevTaxMultiplier = 100;
 
-    event TaxMultiplierSet(uint taxMultiplier);
+    event TaxMultiplierSet(uint256 taxMultiplier);
 
     error InsufficientMEVTax();
 
-    
     /**
      * @dev Set the tax multiplier.
      * @param taxMultiplier The new tax multiplier.
      */
-    function setTaxMultiplier(uint taxMultiplier) public onlyOwner {
+    function setTaxMultiplier(uint256 taxMultiplier) public onlyOwner {
         // todo: check if taxMultiplier is reasonable
 
         // set the tax rate
@@ -31,12 +29,12 @@ abstract contract MEVController is BaseController, MEVTaxedOval {
     }
 
     /**
-     * @notice We use the snapshotData function to pay the MEV tax. 
+     * @notice We use the snapshotData function to pay the MEV tax.
      * @dev This function is called during MEVTaxedOval.unlockLatestValue(), whoever submit this tx with the highest priority fee must pay the MEV tax here.
-     * 
+     *
      */
     function tax() public payable override {
-        uint mevTax = (tx.gasprice - block.basefee) * mevTaxMultiplier;
+        uint256 mevTax = (tx.gasprice - block.basefee) * mevTaxMultiplier;
         if (msg.value < mevTax) revert InsufficientMEVTax();
     }
 }
